+++ server/app.js
@@ -17,7 +17,7 @@ mongoose
 const users = require('./src/routes/users');
 const groups = require('./src/routes/groups');
 const ranking = require('./src/routes/ranking');
-const studying = require('./src/routes/studying');
+const timer = require('./src/routes/timer');
 const board = require('./src/routes/board');
 const comment = require('./src/routes/comment');
 const search = require('./src/routes/search');
@@ -25,7 +25,7 @@ const search = require('./src/routes/search');
 app.use('/api/users', users);
 app.use('/api/groups', groups);
 app.use('/api/ranking', ranking);
-app.use('/api/studying', studying);
+app.use('/api/timer', timer);
 app.use('/api/board', board);
 app.use('/api/comment', comment);
 app.use('/api/search', search);
